Fix the 'no characters in buffer!' panic
Because the X87 80-bit floating point format does not use an extended bit.
It is therefore possible for the significand to be zero.
In from_bits, floating point values with a significand that is zero, but an exponent other than `MAX_EXP - 1` do not get treated as zeroes.
This causes a crash in the formatting code in Display::fmt.
To stick as closely to the LLVM implementation as possible, a workaround in Display::fmt is used to fix this.
When the formatting code detects that the significand is zero, it will format the value as if `category == Category::Zero`.